Rename parseExpirationDate to createExpirationDate
